+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # HackerRank Tour  
+
+## [2017-11-19]  
++ **Added**  
+  - data-structures: linked lists